[move] util package 이동
